Fixed missing resource error
Fixed problem with project not building.  Apparently it was caused by
sprites being deleted from the filesystem that GameMaker then couldn't
find in the project.  Useful knowledge; let's all try not to do that in
the future.  (I'm surprised GM didn't react more negatively.  Crashing
without an explaination is more it's style in these situations.)