Water spread bug squished
It turns out there was a long-standing heisenbug concerning moisture
that, as it turned out, related to bitmath failing because we were
ANDing with a value that turned out to be one bit more than the
(undocumented) maximum value supported by Game Maker.  That fixed,
moisture appears to behave rationally again.  This quirk of Game Maker
has important consequences for insects, BTW: this is probably why bees
act funny, because there's now not enough bits for the bee population to
be held, and so they'll have to be removed or another bug evicted to
make room for them.