Added animal coordinate map
A "map" is GM's term for a hash or dict.  This gives us a faster,
memory-efficent way to determine if a space contains an animal or not.
When we draw animals back into saving and loading, this will have to be
accounted for.  Preferably we'll reconstruct the hash from the incoming
animals, instead of saving the hash wholesale.