Ah ha!  The water can action queue bug is fixed!
The source of the problem turned out to be that, while the code uses
separate symbols for no tool selected ("-1") and the watering can
("TOOL_WATERCAN"), they both resolved to the same value at compile time
(-1).  The problem was solved by creating a new tool, TOOL_NOTOOL,
specifically to cover only the case of no tool being selected.  For the
record, its value is -23.