Bug OOB problem solved (?)
A change was made to coordsinbounds(), a function that's used in many
place to ensure coordinates are good.  This should keep bugs where they
can receive turns and not get stuff off the edge of the map, but because
the function is used in lots of places we should be alert to the
possibility of unexpected errors caused by this.