D and W keys in demomode
These functions should really get their own tools, but I'm in a hurry.
Only while global.demomode is true:
D dries all water on the map by one level
W plops down a lot of water in your current tile (31 units), then
initiates the flow code to find a place for it.  If you're not standing
in a suitable location it'll probably just disappear with effect, so
make sure you're standing in a a basin first.  Note: this destroys
plants.