Greenhouses v2
Ha!  Haha!  Ahhaha!  It works!
You can now plop down a greenhouse just like you can any other building.
You can plant other plants inside of it (so long as they're mostly in
the middle of the greenhouse -- it might be a good idea to provide some
kind of visual feedback of which tiles are legal or not, there are 17
available tiles for planting).